Save the match_id as game_meta, and use it to keep games unique.
POST requests have a "match ID" parameter being sent as an "I" record
within the header. This should be stored in the games table accordingly.
Furthermore, a unique constraint should be placed on the games table
such that all combinations of server_id and match_id are unique in that
table. This prevents duplicate games from being entered by either manual
means or via xonstat-queue.

This change stores the 'I' record in the games table upon submission and
will raise an HTTP 200 exception class if a duplicate entry is found. It
is set to 200 because this is an OK submission, but a duplicate and thus
should not be resubmitted.