Adds an accuracy graph and removes the overall accuracy table.
In retrospect the accuracy/damage table was a cop-out. It doesn't really show anything useful, it was just easy to display query-wise. Meanwhile I've been told time and time again that nex accuracy is the only thing that people are truly looking at as a measure of their accuracy, so that is what this commit does in graphical form. Other useful things will come later, but likely not in a tabular format again.