Make pagination a GET parameter, not part of the URL.
It was not the best idea to add the 'page' displayed to be an actual
part of the URL. It should be a GET parameter instead, which jives
a little better with my understanding of REST-like resources. This
changes all paginated views to use a GET parameter and also limits
results returned to 10 per page for performance and viewability
reasons.