Add a damage efficiency view structure.
Allow tracking of damage efficiency as well. Damage efficiency I'll
define as the amount of damage you do per hit. The higher, the
better! This is appropriate for all splash-damage weapons. Spam shots
are not included on purpose (unless they hit, of course) after
conversations with the competitive community; spam shots usually miss
and are intended for blocking routes or catching your opponent while
they are out of your direct line of sight.