Add elo delta tracking.
An elo delta is the amount by which a player's elo goes up
or down per match. This is useful for tracking purposes
and competitions.

The decimal formatting here is a bit hacky, and I don't like it,
but Mako refuses to do a proper format string on a decimal.Decimal.
This is even thought on the interactive command line it works! Argh.