Remove +/- points display on scoreboard.
I've decided not to show the +/- elo_deltas on the scoreboard
because they are individual measures. They thus belong on a
given player's info page instead. The player_info view is the
future destination of this data (not implemented yet).

Also, I'm preventing duel matches from being processed as both
duel and dm matches. Duel really is a mode of its own, even if
Xonotic doesn't have a dedicated mode for it.