Don't dock Elo points if the player won.
After much debate I've included this enhancement. It prevents
losing Elo points if you've won the game. While this makes the
algorithm not *true* Elo, it prevents the negative perception
in the community. People were turning off their tracking for
fear of losing Elo points even if they won, and this hopefully
will fix that situation.