Get favorite maps for each game type, plus an overall.
The new favorite maps function will get the favorite
maps for each game type. An "overall" type is included
to provide something like the old functionality where
there was an overall favorite map. The implementation
has changed such that the overall favorite map is now the
map you've played the most for any *one* game type; the
different game modes aren't summed beforehand if the maps
are the same between them. I think this is an acceptable
tradeoff given stormkeep in DM is not the same as playing
stormkeep in TDM, mechanically. I may change this in the
future.