Fix broken firefox. Damn you, firefox!
The latest firefox update broke the web borders. This is a
workaround. It compresses the main header by removing the
"get the game" button (not really needed). This change also
swaps out the backgrounds to match xonotic.org and a move of
the title accordingly. All in all I think it looks pretty decent!