Turn off italics in the header.
People didn't like the italics in the header, stating that it
didn't really fit the style. I didn't really have an opinion
either way, so I straightened it back up. I also removed the non-
breaking space after the header that was put in to "even out" the
slant of the italics.