Initial checkin of tabs for the player_info view.
Underneath the player's nick will be a tabbed list of game types,
starting with "overall." Within each of these tabs will be stats
pertaining to only that game type. This allows players to view
fine-grained details about their playing history.

Note that this commit does not style the table much. This is mainly
to ensure that all of the content is being fetched properly.