Make the "recent games" table expand with the "more" link.
Previously a big list of scoreboards would be shown if a person
clicked on the "more" link below the "recent games" table on the
player info page. Now a table that is exactly the same is shown,
but with pagination. This is in essence an attempt to keep the
same elements in the view; when a person clicks on a "more" link,
they expect to see more of the same items they were looking at.

The end benefit for players is that they can quickly page through
all of their games and see summary results as well as their elo
changes from one simple view.