Add a hashkey view similar to the player_info JSON one.
This view is driven by hashkey (not player_id) so the Xonotic
server can query for info using something it knows. The response
value is a JSON object containing the things in the player info
page minus the recent weapons and recent games stuff.

The URL pattern is /hashkey/<hashkey value>.

All this is pretty alpha and is subject to change once we truly
determine our needs  Think of this as a proof of concept for
communicating between the game server and the webapp.