"Intercept" requests during downtime
This is a Golang utility that allows the stats maintainer to still
accept requests from remote servers while the XonStat server is
down for maintenance. It basically takes the requests and stores
them into a small sqlite3 database for future (re)submission.

When users navigate to the normal stats URL, they'll be greeted
with a downpage informing them of maintenance work instead of the
normal app.

Usage of xs_interceptor:
    xs_interceptor [options] <command>

Where <command> is one of the following:
    create   - create the requests db (sqlite3 db file)
    drop     - remove the requests db
    list     - lists the requests in the db
    serve    - listens for stats requests, storing them if found
    resubmit - resubmits the requests to another URL

Where [options] is one or more of the following:
    -port    - port number (int) to listen on for 'serve' command
    -url     - url (string) to submit requests