Separate scoreboards by team.
This is in preparation for the display of team scores. The team
score will be shown to the left of the scoreboard. The tricky
thing with this implementation is keeping the columns aligned
properly. For this I've added a player-nick class (again) as a
percentage to make the two (or more) tables line up.