Initial stacked area chart for damage.
This version doesn't work. It is wanting an entry for every
weapon in every game, and I don't have that yet. I have to loop
over all the weapons present in the set thus far and then create
zero-ed out objects for the games in which certain weapons don't
exist. Sigh - I haven't decided if I want to do this in the view
or in Javascript.