Add links to the game modes in the stat line.
The implementation of this is a bit funky. See the
comment in main.py for the get_day_summary_stats
function call. Long story short is I picked what I
believe to be the lesser of two evils - I'm
assembling a complex string inside the view instead
of putting a Python code block directly in the template.