Speed up weaponstats requests by 2.5x.
We can exploit the GIN index on games.players to quickly get a list of the most
recently-played games. We construct this list as a subquery that can be used in
an IN list in the main weaponstats query, greatly reducing the time needed to
fetch the rows. It's about 2.5 times faster according to apache bench.
Additionally this query is helped by moving it to a subquery, whereas once it
was two queries (thus two round trips to the DB).