Remove the join to game stats for accuracy data.
This speeds up the game_info view considerably. On my local machine the
requests go from .62 to about .2 seconds! I believe the pgstat rows were
included in the response to show nicknames in the accuracy table.

I'll have to do further checks to ensure that if a pgstat row exists a
corresponding pwstat row exists as well.